requirements.txt
asgiref==3.4.1
Django==3.2.5
django-background-tasks==1.2.5
django-cleanup==5.2.0
django-compat==1.0.15
Pillow==8.3.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
